Rename conflicting files in `platform/win`
