+++ source/InjectHooksMain.cpp
@@ -493,7 +493,7 @@
 
 #include "platform/win/VideoPlayer/VideoPlayer.h"
 #include "platform/win/VideoMode.h"
-#include "platform/win/Platform.h"
+#include "platform/win/WinPlatform.h"
 
 #include "app/app.h"
 #include <RealTimeShadowManager.h>
