+++ source/app/platform/win/WinPlatform.h
